export character symbols
